Convert files to ESM
