+++ test/video.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-useless-escape */
-const test = require('ava').default;
-const MarkdownItRenderer = require('../src');
+import test from 'ava';
+import MarkdownItRenderer from '../src/index.js';
 
 test('MarkdownItRenderer.render(content, config): can render a video', (t) => {
   const markdown = '<video src="/uploads/example.mp4">';
