+++ test/youtube.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-useless-escape */
-const test = require('ava').default;
-const MarkdownItRenderer = require('../src');
+import test from 'ava';
+import MarkdownItRenderer from '../src/index.js';
 
 test('MarkdownItRenderer.render(content, config): can render a YouTube video', (t) => {
   const markdown = '<youtube v="aR3fVuLEtj8" width="560" height="315" title="YouTube Video Player" start="0">';
